consumer/requirements.txt
faust-streaming==0.10.14
asyncpg==0.29.0
python-dotenv==1.0.1
